Prevent gin-bin from being created on host
gin automatically compiles to a binary with filename gin-bin, if we are
running gin from inside the mounted volume, this will get written out to
the host.  We can instead run gin from a working directory outside the
volume and pass it a path to the files.

This also now allows us to mount the volume as read-only.