Use kong image that has better debug logging
The official kong image does not log errors to stdout, which is
annoying when developing in docker compose.

Also removed the "restart: always" as it was tarting up kong even
after I killed it via Ctrl+C and restarted my computer.  This
messed up my network settings as the container kept starting and
crashing.